requirements.txt
flask
sqlalchemy
mysql-connector-python
python-dotenv
PyMySQL